分析/requirements/base.txt
# Base dependencies
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
redis==4.5.4
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
jieba==0.42.1
scikit-learn==1.3.0
